[SR-1058] Extend aliases in protocols to be included in ArchetypeType.
Previously I thought I could get away with the aliases only being in
the PotentialArchetypes, but with inherited protocols we can get asked
for the dependent type (of the aliases name) of a base ArchetypeType
and there’s no way to get the ‘destination’ of the alias unless it’s
already a NestedType of the ArchetypeType.

So now there’s an additional flag bit on NestedType for whether this
archetype is an alias (meaning it is for lookups only and shouldn’t be
a separate generic parameter) and there’s some extra machinery for
making sure that all aliases get recorded as NestedTypes in the
archetypes.

Before this gets pulled in it needs more tests, and the alias flag bit
needs to be de/serializable.