AngryPhonebook: uppercased and lowercased are functions so call them
We were not doing any work in this benchmark at -O because we were removing dead
code (to build a tuple of partially applied functions). Now that we are actually
calling the functions we can't eliminate dead code.

This is going to appear as a regression but it is none.