Enablement of ObjectiveCBridgeable protocol for Linux
Make the ObjectiveCBridgeable protocol available
and minimal pieces of the runtime support for it
available even without an Objective-C runtime.
This is motivated by supporting bridging in
swift-corelibs-foundation on Linux via the
compiler/runtime support provided by
ObjectiveCBridgeable.

Also enable the six existing testcases for
ObjectiveCBridgeable that don't import Foundation
or truly rely on an Objective-C runtime being present.