Work around relocation R_X86_64_PC32 link error
A more complete discussion of the error and its background can be found
in SR-1023.

The brief overview is that in versions of `ld` that would be effected by
this dynamic relocation issue there are two flags: `-z nocopyreloc` and
`-z noextern-protected-data`. If `ld` supports the flags then they are
sento to the linker. This patch checks that the linker supports the
flags and then optionally sets them.

The code to check if `ld` supports the flags is inspired by this CMake
mailing list entry.

https://cmake.org/pipermail/cmake/2011-July/045525.html