Add a new utility called 80+-check.
This utility takes input from stdin and prints out any lines that are longer
than 80+ lines. If it finds any such violations it prints out the line number,
the length of the line, and a version of the line with the left and right sides
of the line stripped of whitespace. The stripping is done so that useful output
is provided even when a line has a large amount of whitespace (something that
can happen when copy/pasting in terminal).

If any violation is found while processing the lines of the file, the script
fails. This happens only after all lines have been processed.