stdlib: Rewrite _getClassPlaygroundQuickLook() in Swift
The C++ code was very fragile in terms of ABI dependencies and broke
when the standard library was built with -enable-resilience.

The actual reason it broke is that case numbering changes when
resilience is enabled, but instead of messing with that, it seemed
more logical to rewrite this routine in Swift instead, to avoid
ABI dependencies altogether.

This requires using the "shadow protocol" trick to call NSNumber
methods, since we cannot import NSNumber from the stdlib.