[Definite initialization] Recognize delegation to a C function imported as an initializer.
Add yet one more flavor of hack to DI to recognize where we are
delegating the initialization of 'self'. The existing hack in this
area (for Objective-C factory initializers) is based on recognizing
the value_metatype instruction that feeds into the application of the
factory initializer. C functions imported as initializers don't have a
metatype argument, so instead tag the assignment into the self box as
the initialization of self.

As a minor cleanup in this area, don't emit the dead value_metatype
instruction when invoking a C-imported factory initializer.