[build-script] Don't include revision numbers in --version by default.
Previously, any commit or rebase would force a re-link of the compiler to pick up
the updated version number, even if nothing in the compiler changed. That in turn
forces an unnecessary rebuild of the standard library.

Now, you can opt into this behavior with the build-script-impl option
'swift-include-revisions-in-version' (it's off by default). When not turned on,
no revision numbers will be shown in --version output.  As far as the standard
presets go, this is turned on for non-incremental buildbot builds and package
builds, and off in incremental buildbot builds.