Only check the minimal set of generic requirements when opening a generic function type during constraint solving, as opposed to checking a bunch of implicit things that we already know.  This should significantly improve the efficiency of checking uses of generic APIs by reducing the total number of type variables and constraints.
It is becoming increasingly funny to refer to this minimized generic
signature as the "mangling" signature.

The test changes are kind of a wash: in one case, we've eliminated
a confusing extra error, but in another we've caused the confusing
extra error to refer to '<<error type>>'.  Not worth fighting right
now.  The reference-dependencies change is due to not needing to
pull in all of those associated types anymore, which seems correct.