[docs] Clarify SIL-level TBAA rules.
This documentation now reflects the optimizer's reality.

I'm still working on formalizing language level rules for strict aliasing. Those
will be introduced in a separate doc file.

There some language about alias-introducing operations. This concept existed in
the old documentation but was never really implemented. This all makes sense now
that we have a formal model for binding memory to a type along with specific variants
of pointer_to_address that either enforce strict aliasing or permit type punning.

The detailed explanation of TBAA should probably be moved into a separate
optimizer document, but there isn't a good place for it yet.