[Sema] Warn unhanded errors in empty catch block
Implementation for https://bugs.swift.org/browse/SR-978
A warning is emitted when the only `catch` on a `do/catch` statement is
unguarded and has no body. A fixit is also offered to remove the catch
block. The tests were updated to fix new warnings and to add cases for
this new functionality.