[WIP] Refactor FSO
Several functionalities have been added to FSO over time and the logic has become
muddled.

Refactor 1 single function to do all sorts of different transformations in FSO into several
small transformations, each of which does a specific job.

Eventually, we will have this infrastructure to write different transformations of the signature
of a function.

And a transformation only needs to define 2 functions, transform - how to transform the function.
rewriteCallSite - hwo to rewrite callsite to make calls to the transformed function.