ARC black magic for _NSContiguousString.getCharacters
What happened here is that we owned to guaranteed specialize getCharacters but
then did not inline the specialized function leaving us with the stub:

  getCharacters() {
    retain self
    getCharacters_o2g(self, ...)
    release self
  }

If we mark the function with always inline getCharacters_o2g will get inlined
and the retain/release is removed.

This helps hashing and string comparison and should consequently speedup
Dictionary and Set of String.

rdar://25797071