Relax verification of dealloc_stack.
Allow it to have undef as an operand.

This can happen when NoReturnFolding does RAUW for the instructions that
come after a @noreturn function, replacing the uses of those
instructions in blocks that are unreachable. These instructions end up
getting deleted during diagnose unreachable when we remove the
unreachable code.

Fixes SR-967 / rdar://problem/25882880.