Revert "[Preset] Dont skip lldb test on Linux package preset"
This reverts commit 42357b67b23a8c5bfe7fd7f76e54047cefbac6d7.

On Ubuntu, the LLDB REPL tests fail when using the LLDB from
the build dir as the LLDB test suite uses it.  Those tests
pass when using the install layout built into {root}/usr/bin/lldb.

I'll re-enable this once I have that investigated and working
properly.

This should fix the new issues on the Ubuntu package CI for
the Swift master branch.