[test] Update reference-dependencies.swift.
There's no longer anything looking up any conformances on Int and
getting negative results. Previously this was in looking for viable
'~=' overloads for the 'case' on line 206, but that seems to be
ruled out by other means now. We still have CHECK lines for other
protocol conformance checks, so we're not losing coverage here.

rdar://problem/25890224