Reflection: Demangling for lowered function types
Lowered function types appear in capture descriptors.

For now, we completely punt on demangling parameters and
results, and just hackishly map the SIL function type
representation to an AST function type representation.

This is good enough, because all we care about is whether
we have a raw function pointer, raw function pointer with
context, or a block.