[CMake] Allow *only* static stdlib to be built
Prior to these changes, it was only possible to build static libraries
for the standard library and SDK overlays if *also* building dynamic
libraries. This is problematic if, for example, you wanted to defer the
linking of the standard library build products until later.

These changes allow you to build *only* static libraries, using
something like the following command:

```
$ utils/build-script -R --skip-build-benchmarks \
    -- --build-swift-static-stdlib=1 --build-swift-dynamic-stdlib=0
```